refactored the grading folder
