+++ src/main/java/autocadDrawingChecker/grading/GradedExport.java
@@ -1,5 +1,6 @@
 package autocadDrawingChecker.grading;
 
+import autocadDrawingChecker.grading.criteria.AbstractGradingCriteria;
 import autocadDrawingChecker.data.elements.AutoCADExport;
 import java.util.HashMap;
 import java.util.List;
