+++ src/main/java/autocadDrawingChecker/grading/criteria/implementations/LineStart.java
@@ -1,28 +1,28 @@
-package autocadDrawingChecker.grading.implementations;
+package autocadDrawingChecker.grading.criteria.implementations;
 
 import autocadDrawingChecker.data.elements.AutoCADLine;
 import autocadDrawingChecker.data.elements.AutoCADElement;
 
 /**
  *
- * @author Matt
+ * @author Matt Crow
  */
-public class LineEnd implements AbstractVectorCriteria<AutoCADLine> {
+public class LineStart implements AbstractVectorCriteria<AutoCADLine> {
     @Override
     public String getDescription() {
-        return "Grades based on how closesly the student's line end points match up with those of the instructor's";
+        return "Grades based on how closesly the student's line start points match up with those of the instructor's";
     }
 
     @Override
     public String getName() {
-        return "Line End";
+        return "Line Start";
     }
 
     @Override
     public double[] extractVector(AutoCADLine e) {
         double[] v = new double[AutoCADLine.DIMENSION_COUNT];
         for(int i = 0; i < v.length; i++){
-            v[i] = e.getEnd(i);
+            v[i] = e.getStart(i);
         }
         return v;
     }
