+++ src/main/java/autocadDrawingChecker/start/DrawingCheckerProperties.java
@@ -1,6 +1,6 @@
 package autocadDrawingChecker.start;
 
-import autocadDrawingChecker.grading.AbstractGradingCriteria;
+import autocadDrawingChecker.grading.criteria.AbstractGradingCriteria;
 import java.util.List;
 import java.util.Properties;
